+++ students/k33401/Kobelev_Lev/Practical_Works/simple_django_web_project/practical_work_1/apps.py
@@ -3,4 +3,4 @@ from django.apps import AppConfig
 
 class Task12Config(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "task_1_2"
+    name = "practical_work_1"
