+++ students/k33401/Kobelev_Lev/Practical_Works/simple_django_web_project/practical_work_1/models.py
