+++ students/k33401/Kobelev_Lev/Practical_Works/simple_django_web_project/simple_django_web_project/settings.py
@@ -31,7 +31,7 @@ ALLOWED_HOSTS = []
 # Application definition
 
 INSTALLED_APPS = [
-    "task_1_2",
+    "practical_work_1",
     "django.contrib.admin",
     "django.contrib.auth",
     "django.contrib.contenttypes",
